producer-service/requirements.txt
# FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
starlette==0.27.0

# Kafka client
kafka-python==2.0.2

# Prometheus metrics
prometheus-client==0.19.0

# HTTP client for health checks
httpx==0.25.2

# Environment variable management
python-dotenv==1.0.0

# Data validation and serialization
pydantic==2.5.0

# Logging and monitoring
structlog==23.2.0

# Development and testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# TODO: Add OpenTelemetry for distributed tracing
# opentelemetry-api==1.21.0
# opentelemetry-sdk==1.21.0
# opentelemetry-instrumentation-fastapi==0.42b0
# opentelemetry-instrumentation-kafka==0.42b0

# TODO: Add Redis for caching and session management
# redis==5.0.1

# TODO: Add database dependencies if needed
# sqlalchemy==2.0.23
# alembic==1.12.1 